// src/app/interfaces/tecnologia.interface.ts
// Generated by https://quicktype.io

export interface ITecnologia {
    status:       string;
    totalResults: number;
    results:      Result[];
    nextPage:     string;
}

export interface Result {
    title:       string;
    link:        string;
    keywords:    string[];
    creator:     string[] | null;
    video_url:   null;
    description: string;
    content:     null | string;
    pubDate:     string;
    image_url:   null | string;
    source_id:   string;
    category:    Category[];
    country:     Country[];
    language:    Language;
    expanded?: boolean;
}

export enum Category {
    Technology = "technology",
}

export enum Country {
    Spain = "spain",
}

export enum Language {
    Spanish = "spanish",
}